Receiving classes added.

+++ IrSenderNonMod.cpp
@@ -20,8 +20,8 @@ this program. If not, see http://www.gnu.org/licenses/.
 IrSenderNonMod::IrSenderNonMod(pin_t pin) : IrSender(pin) {
 }
 
-void IrSenderNonMod::send(const microseconds_t buf[], uint16_t len) {
-    for (uint16_t i = 0; i < len; i++) {
+void IrSenderNonMod::send(const microseconds_t buf[], unsigned int len) {
+    for (unsigned int i = 0; i < len; i++) {
         digitalWrite(outputPin, (i & 1) ? LOW : HIGH);
         delayUSecs(buf[i]);
     }
